requirements.txt
Django==3.1
djangorestframework==3.12
drf-yasg==1.20.0
django-mysql==4.4.0
django-cors-headers==3.11.0
mysqlclient==2.1.0
pika==1.2.0
pyyaml==6.0
